+++ config/plugins/render.js
@@ -18,7 +18,8 @@ loginUrl += loginUrl.indexOf('?') == -1 ? '?' : '&';
 // 渲染模板输出
 ctrlUtil.prototype.renderView = function(view, data) {
     var req = this.get('req'),
-        res = this.get('res');
+        res = this.get('res'),
+        errorCallback;
 
     // 输出错误信息到页面，不渲染模板
     if (data.err_no != 0) {
@@ -27,11 +28,23 @@ ctrlUtil.prototype.renderView = function(view, data) {
             this.sessionDestroy();
             return;
         }
+
+        if (!data.err_msg) {
+            data.err_msg = defaultMsg;
+        };
+
         // gearman请求出错
         if (!data.IGNORE_ERR) {
-            // 抛出错误
-            req.next(new Error(data.err_msg || defaultMsg));
-            return;
+            errorCallback = req.onRenderViewError || this.onRenderViewError;
+            // 默认的模板错误
+            if (!errorCallback) {
+                // 应用返回的错误，暂时直接输出到页面上
+                // TODO: 错误处理的约定及输出方式的优化
+                res.send(data.err_msg);
+                // res.next(new Error(data.err_msg));
+            } else {
+                errorCallback.apply(this, arguments);
+            }
         }
     }
     return renderView.apply(this, arguments);